add storage (#40)
* add storage

* storage

* python2 compatible

* db store

* provide doc

* rename example to examples

* add expiration

* fix a bug

* more dov

* bug fixes

* decorator

* storage

* storage

* storage
